refactor: remove state from ListBox and Option and export SelectContext
